+++ config/permissions.yaml
@@ -45,9 +45,13 @@
     - 100
 "ajax.changepassword":
     - 21
+"ajax.changeownpassword":
+    - 100
 "ajax.planningjobcheckcopy":
     - 100
 "ajax.mailtest":
+    - 100
+"ajax.checkpassword":
     - 100
 "ajax.update_agent_login":
     - 21
