+++ templates/agents/agent_table_edit.html.twig
@@ -237,30 +237,7 @@
   {% endif %}
 
   {% if not demo %}
-    <div id="reset-password-form" title="Réinitialisation du mot de passe" class='noprint' style='display:none;'>
-      <form>
-        <div>
-          <label for="new_password">Nouveau mot de passe</label>
-          <input class="password-input" type='password' id="new_password" name="new_password"/>
-        </div>
-
-        <br />
-
-        <div>
-          <label for="confirm_new_password">Confirmez le nouveau mot de passe</label>
-          <input class="password-input" type='password' id="confirm_new_password" name="confirm_new_password"/>
-        </div>
-
-        <br />
-
-        <div class="password-error" id="password-dont-match">
-            Les mots de passe ne sont pas identiques.
-        </div>
-        <div class="password-error" id="password-empty">
-            Le mot de passe ne peut pas être vide.
-        </div>
-      </form>
-    </div>
+    {% include 'agents/elements/change_password_div.html.twig' %}
 
     <div id="reset-login-form" title="Réinitialisation du login" class='noprint' style='display:none;'>
       <form>
