+++ templates/agents/elements/change_password.html.twig
@@ -1,0 +1,92 @@
+{# agents/elements/change_password.html.twig #}
+  <style>
+    .password-input {
+      position: absolute;
+      left: 300px;
+    }
+    .password-error {
+      color: red;
+      display: none;
+    }
+  </style>
+  <script type='text/JavaScript'>
+    function save_password(password) {
+      {% if get_agent_id == true %}
+        var agent_id = $('input[name="id"]').val();
+      {% endif %}
+      $.ajax({
+        url: url('{{ change_password_url }}'),
+        type: "POST",
+        dataType: "html",
+        {% if get_agent_id == true %}
+            data: {password: password, id: agent_id},
+        {% else %}
+            data: {password: password},
+        {% endif %}
+        success: function(){
+            $("#reset-password-form").dialog( "close" );
+        },
+        error: function(xhr, ajaxOptions, thrownError){
+          alert("Erreur lors de la modification du mot de passe");
+        }
+      });
+    }
+
+    $(document).ready(function(){
+      $("#reset-password-form").dialog({
+        autoOpen: false,
+        height: 280,
+        width: 580,
+        modal: true,
+        buttons: {
+          Enregistrer: function() {
+            $('#password-dont-match').hide();
+            $('#password-empty').hide();
+            $('#password-too-weak').hide();
+            $('#password-check-error').hide();
+
+            password = $('#new_password').val();
+            confirm_password = $('#confirm_new_password').val();
+            password_check = false;
+
+            $.ajax({
+                url: url('ajax/check-password'),
+                type: "POST",
+                dataType: "html",
+                async: false,
+                data: {password: password},
+                success: function(result){
+                    if (result == "ok") {
+                        password_check = true;
+                    } else {
+                        $('#password-too-weak').show();
+                    }
+                },
+                error: function(xhr, ajaxOptions, thrownError){
+                  $('#password-check-error').show();
+                }
+              });
+
+
+            if (password != confirm_password) {
+              $('#password-dont-match').show();
+            }
+            else if (password == '') {
+              $('#password-empty').show();
+            } else if (password_check == true) {
+              save_password(password);
+            }
+          },
+          Annuler: function() {
+            $(this).dialog( "close" );
+          },
+        },
+      });
+
+      $('#change-password-link').on('click', function() {
+        $("#reset-password-form").dialog( "open" );
+      });
+
+    });
+  </script>
+
