+++ templates/agents/elements/change_password_div.html.twig
@@ -1,0 +1,33 @@
+{# agents/elements/change_password_div.html.twig #}
+    <div id="reset-password-form" title="Réinitialisation du mot de passe" class='noprint' style='display:none;'>
+      <form>
+        <div>
+          <label for="new_password">Nouveau mot de passe</label>
+          <input class="password-input" type='password' id="new_password" name="new_password"/>
+        </div>
+
+        <br />
+
+        <div>
+          <label for="confirm_new_password">Confirmez le nouveau mot de passe</label>
+          <input class="password-input" type='password' id="confirm_new_password" name="confirm_new_password"/>
+        </div>
+
+        <br />
+
+        <div class="password-error" id="password-dont-match">
+            Les mots de passe ne sont pas identiques.
+        </div>
+        <div class="password-error" id="password-empty">
+            Le mot de passe ne peut pas être vide.
+        </div>
+        <div class="password-error" id="password-too-weak">
+            Le mot de passe est trop faible. Le mot de passe doit faire au minimum 8 caractères, et contenir au moins une minuscule, une majuscule, un chiffre et un caractère spécial.
+            Les caractères spéciaux sont les suivants: !"#$%&'()*+,-./:;<=>?@[\]^_`{|}~
+        </div>
+        <div class="password-error" id="password-check-error">
+            Impossible de vérifier la robustesse du mot de passe. Veuillez contacter votre administrateur.
+        </div>
+      </form>
+    </div>
+
